cleanup and delay
