+++ modules/core/src/main/scala/gem/config/InstrumentConfig.scala
@@ -1,6 +1,20 @@
 package gem
 package config
 
-import gem.enum.Instrument
+import gem.enum._
 
-class InstrumentConfig(val instrument: Instrument)
+import java.time.Duration
+
+sealed abstract class InstrumentConfig extends Product with Serializable
+
+final case class F2Config(
+  fpu: F2FpUnit,
+  mosPreimaging: Boolean,
+  exposureTime: Duration,
+  F2filter: F2Filter,
+  lyoutWheel: F2LyotWheel,
+  disperser: F2Disperser
+) extends InstrumentConfig
+
+
+
