+++ modules/core/src/main/scala/gem/config/SmartCalConfig.scala
@@ -1,11 +1,6 @@
 package gem
 package config
 
-import gem.describe._
-import gem.describe.Metadata.Access.Science
-import gem.describe.Metadata.Scope.SingleStep
-import gem.describe.Metadata.{Attrs, Label}
-
 import scalaz._
 
 final case class SmartCalConfig(smartCalType: SmartCalConfig.Type)
@@ -21,21 +16,5 @@ object SmartCalConfig {
     val Default = Arc
     val All     = NonEmptyList(Arc, Flat, DayBaseline, NightBaseline)
   }
-
-  val Lab = Label("Smart")
-
-  object TypeProp extends Prop[SmartCalConfig] {
-    type B = Type
-    val eq: Equal[Type] = Equal.equalA
-
-    def lens: SmartCalConfig @> Type = Lens.lensu((a, b) => a.copy(smartCalType = b), _.smartCalType)
-
-    val meta = EnumMetadata[Type](Attrs(Label(Lab, "Cal Type"), Science, SingleStep), Type.All)
-  }
-
-  implicit val DescribeSmartCal: Describe[SmartCalConfig] =
-    Describe.forProps(
-      SmartCalConfig(SmartCalConfig.Type.Default),
-      TypeProp
-    )
 }
+
