+++ modules/describe/src/main/scala/gem/describe/Prop.scala
